* ./hurricane/doc/hurricane/src/hurricane/Layer.{h.cpp}    ./hurricane/doc/hurricane/src/hurricane/BasicLayer.{h.cpp}    - Changed "Type" to "Material" to fit better the OpenAccess semantic.        Introduce Material meanings : pWell, nWell, pImplant, nImplant        active, metal, cut, obstacle & other.
 * ./coriolis/src/crlcore :
   - Switched to the generic layer model. Updated technology XML configuration
       files, technology parser & AP parser/driver.
   - Note that the representation of transistors visually changes as it's now
       modelled with Implant + Active.

+++ hurricane/src/hinspector/HInspectorWidget.cpp
@@ -115,7 +115,7 @@ void HInspectorWidget::keyPressEvent(QKeyEvent *event) {
     }
 }
 
-void HInspectorWidget::recordChanged(int index) {
+void HInspectorWidget::recordChanged(size_t index) {
     if (index >= 0 && index < filterProxyModelsHistory.size()) {
         QSortFilterProxyModel* proxyModel = filterProxyModelsHistory[index];
         slotsView->setModel(proxyModel);
